Keep backend joint-drive hook import free of USD libraries
The PhysX schema package registered its joint-drive skip predicate by
importing the register function from the core schemas module at package
import time. That core module imports pxr at module top, so importing the
backend package eagerly pulled USD libraries into otherwise USD-free import
paths (e.g. resolving a robot config), failing the omni/pxr-free import
guard.

Move the registry and registration API into a dedicated USD-free hooks
module so a backend can register its hook at import time without dragging
in pxr.
